# Conflicts:
#	SP4/Base.v12.suo
#	SP4/Base/Debug/Base.log
#	SP4/Base/Lua/PlayerSave3.lua
#	SP4/Common/Debug/Common.log

+++ SP4/Base/Source/Main/Game/Scene/Town.h
@@ -22,7 +22,7 @@ private:
 	bool OpenSetting;
 	bool OpenMission;
 	Camera camera;
-	int Volume;
+	float Volume;
 
 public: 
 	Town();
